agregado funcionalidad de borrado de comentarios solo para administradores en pantalla mascotas, refugios y cuidadores

+++ src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosComentario.java
@@ -4,10 +4,10 @@ public class DatosComentario {
     private int clasificacion;
     private String mensaje;
     private int idUsuario;
-    private Integer idCuidado;
-    private Integer idMascota;
+    private Long idCuidado;
+    private Long idMascota;
 
-    private Integer idComentarioPadre;
+    private Long idComentarioPadre;
 
     public void setClasificacion(int clasificacion) {
         this.clasificacion = clasificacion;
@@ -33,27 +33,27 @@ public class DatosComentario {
         this.idUsuario = idUsuario;
     }
 
-    public void setIdCuidado(Integer idCuidado) {
+    public void setIdCuidado(Long idCuidado) {
         this.idCuidado = idCuidado;
     }
 
-    public void setIdMascota(Integer idMascota) {
+    public void setIdMascota(Long idMascota) {
         this.idMascota = idMascota;
     }
 
-    public Integer getIdComentarioPadre() {
+    public Long getIdComentarioPadre() {
         return idComentarioPadre;
     }
 
-    public void setIdComentarioPadre(Integer idComentarioPadre) {
+    public void setIdComentarioPadre(Long idComentarioPadre) {
         this.idComentarioPadre = idComentarioPadre;
     }
 
-    public Integer getIdCuidado() {
+    public Long getIdCuidado() {
         return idCuidado;
     }
 
-    public Integer getIdMascota() {
+    public Long getIdMascota() {
         return idMascota;
     }
 }
