+++ src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/ServicioUsuario.java
@@ -8,5 +8,5 @@ public interface ServicioUsuario {
 
 	String encriptarClave(String clave);
 
-	Usuario consultarUsuario(int id);
+	Usuario consultarUsuario(long id);
 }
